add new fixtures for tests. Fix endline character for files downloaded from Github.

+++ tests/conftest.py
@@ -24,12 +24,27 @@ def download_raw_test_data():
         "centroid-data-zlib-no-index-compressed.mzML",
         "profile-data-zlib-indexed-compressed.mzML"
     ]
+
+    WINDOWS_LINE_ENDING = b'\r\n'
+    UNIX_LINE_ENDING = b'\n'
+
     for f in files:
         file_url = dataset_url + "/" + f
         r = requests.get(file_url)
         file_path = os.path.join(dataset_path, f)
-        with open(file_path, "w") as fin:
-            fin.write(r.text)
+        if not os.path.isfile(file_path):
+            # save content
+            with open(file_path, "w") as fin:
+                fin.write(r.text)
+
+            # convert end line characters for windows
+            if os.name == "nt":
+                with open(file_path, "rb") as fin:
+                    content = fin.read()
+                content = content.replace(WINDOWS_LINE_ENDING, UNIX_LINE_ENDING)
+
+                with open(file_path, "wb") as fin:
+                    fin.write(content)
 
 
 # simulated data used for tests
@@ -53,6 +68,21 @@ def data_container_with_order():
 
 
 @pytest.fixture
+def data_container_with_order_single_qc():
+    population = {"healthy": 20, "disease": 35}
+    mean = {"healthy": np.array([50, 100, 150]),
+            "disease": np.array([150, 200, 300])}
+    cov = {"healthy": np.array([1, 1, 1]),
+           "disease": np.array([2, 2, 2])}
+    blank_contribution = np.array([3, 5, 10])
+    mz = np.array([100, 200, 300])
+    rt = np.array([50, 60, 70])
+    dc = simulate_dataset(population, mean, cov, mz, rt, blank_contribution,
+                          prepend_blank=1, append_blank=1, triple_qc=False)
+    return dc
+
+
+@pytest.fixture
 def data_container_with_batch_effect():
     population = {"healthy": 24, "disease": 24}
     mean = {"healthy": np.array([50, 100, 150]),
@@ -72,6 +102,8 @@ def data_container_with_batch_effect():
         return x.multiply(batch_effect, axis=0)
     dc._data_matrix = (dc.data_matrix.groupby(dc.batch)
                        .apply(add_batch_effect))
+    mapping = {"sample": list(population), "qc": ["QC"]}
+    dc.mapping = mapping
     return dc
